Move _function_key as well
